refactor: recipes details fetching logic inside useRecipeDetails
