+++ frontend/src/context/recipes-context.tsx
@@ -5,6 +5,7 @@ import {
     useContext,
     useMemo,
 } from 'react'
+import useRecipeDetails from '../hooks/useRecipeDetails'
 
 export const RecipesContext = createContext<{
     cuisines: Record<number, string>
@@ -23,31 +24,15 @@ export const useRecipesContext = () => {
 }
 
 const RecipesProvider: FC<PropsWithChildren> = ({ children }) => {
-    const { cuisines, diets, difficulties } = {} as {
-        cuisines: Record<'id' | 'name', string>[]
-        diets: Record<'id' | 'name', string>[]
-        difficulties: Record<'id' | 'name', string>[]
-    }
-
-    const cuisinesMap = Object.fromEntries(
-        cuisines.map((obj) => [obj.id, obj.name])
-    )
-
-    const dietsMap = Object.fromEntries(diets.map((obj) => [obj.id, obj.name]))
-
-    const difficultiesMap = Object.fromEntries(
-        difficulties.map((obj) => [obj.id, obj.name])
-    )
-
-    console.log(cuisinesMap, dietsMap, difficultiesMap)
+    const { cuisines, diets, difficulties } = useRecipeDetails()
 
     const recipesCtx = useMemo(() => {
         return {
-            cuisines: cuisinesMap,
-            diets: dietsMap,
-            difficulties: difficultiesMap,
+            cuisines,
+            diets,
+            difficulties,
         }
-    }, [cuisinesMap, dietsMap, difficultiesMap])
+    }, [cuisines, diets, difficulties])
 
     return (
         <RecipesContext.Provider value={recipesCtx}>
